+++ src/main/java/com/traveltime/sdk/dto/common/transportationfast/Cycling.java
@@ -1,0 +1,6 @@
+package com.traveltime.sdk.dto.common.transportationfast;
+
+import lombok.Value;
+
+@Value
+public class Cycling implements Transportation { }
