+++ src/main/java/com/traveltime/sdk/dto/common/transportationfast/CyclingAndFerry.java
@@ -1,0 +1,6 @@
+package com.traveltime.sdk.dto.common.transportationfast;
+
+import lombok.Value;
+
+@Value
+public class CyclingAndFerry implements Transportation { }
