+++ src/main/java/com/traveltime/sdk/dto/common/transportationfast/PublicTransport.java
@@ -1,0 +1,6 @@
+package com.traveltime.sdk.dto.common.transportationfast;
+
+import lombok.Value;
+
+@Value
+public class PublicTransport implements Transportation { }
