+++ src/main/java/com/traveltime/sdk/dto/requests/timefilterfast/ManyToOne.java
@@ -1,6 +1,7 @@
 package com.traveltime.sdk.dto.requests.timefilterfast;
 
 import com.traveltime.sdk.dto.common.Property;
+import com.traveltime.sdk.dto.common.transportationfast.Transportation;
 import lombok.*;
 import lombok.extern.jackson.Jacksonized;
