--- src/main/java/com/traveltime/sdk/dto/requests/timefilterfast/transportation/PublicTransport.java
@@ -1,7 +1,0 @@
-package com.traveltime.sdk.dto.requests.timefilterfast.transportation;
-
-import com.traveltime.sdk.dto.requests.timefilterfast.Transportation;
-import lombok.Value;
-
-@Value
-public class PublicTransport implements Transportation { }
