+++ src/main/java/com/traveltime/sdk/dto/requests/timemapfast/ManyToOne.java
@@ -1,6 +1,7 @@
 package com.traveltime.sdk.dto.requests.timemapfast;
 
 import com.traveltime.sdk.dto.common.Coordinates;
+import com.traveltime.sdk.dto.common.transportationfast.Transportation;
 import com.traveltime.sdk.dto.requests.timemapfast.levelofdetail.LevelOfDetail;
 import lombok.*;
 import lombok.extern.jackson.Jacksonized;
