--- src/main/java/com/traveltime/sdk/dto/requests/timemapfast/transportation/DrivingAndPublicTransport.java
@@ -1,8 +1,0 @@
-package com.traveltime.sdk.dto.requests.timemapfast.transportation;
-
-import com.traveltime.sdk.dto.requests.timemapfast.Transportation;
-import lombok.Value;
-
-@Value
-public class DrivingAndPublicTransport implements Transportation {
-}
