--- src/main/java/com/traveltime/sdk/dto/requests/timemapfast/transportation/Walking.java
@@ -1,7 +1,0 @@
-package com.traveltime.sdk.dto.requests.timemapfast.transportation;
-
-import com.traveltime.sdk.dto.requests.timemapfast.Transportation;
-import lombok.Value;
-
-@Value
-public class Walking implements Transportation { }
