+++ src/test/java/com/traveltime/sdk/TimeFilterTest.java
@@ -6,13 +6,13 @@ import com.traveltime.sdk.dto.common.FullRange;
 import com.traveltime.sdk.dto.common.Location;
 import com.traveltime.sdk.dto.common.Property;
 import com.traveltime.sdk.dto.common.transportation.PublicTransport;
+import com.traveltime.sdk.dto.common.transportationfast.DrivingAndPublicTransport;
 import com.traveltime.sdk.dto.requests.TimeFilterFastRequest;
 import com.traveltime.sdk.dto.requests.TimeFilterRequest;
 import com.traveltime.sdk.dto.requests.timefilter.*;
 import com.traveltime.sdk.dto.requests.timefilterfast.ArrivalSearches;
 import com.traveltime.sdk.dto.requests.timefilterfast.ManyToOne;
 import com.traveltime.sdk.dto.requests.timefilterfast.OneToMany;
-import com.traveltime.sdk.dto.requests.timefilterfast.transportation.DrivingAndPublicTransport;
 import com.traveltime.sdk.dto.responses.*;
 import com.traveltime.sdk.dto.responses.errors.TravelTimeError;
 import io.vavr.control.Either;
